+++ e2m2e/integrators.py
@@ -23,6 +23,7 @@ if TYPE_CHECKING:
     batch_synodic_to_j2000_py: Any
     build_cr3bp_hamiltonian_py: Any
     check_collision_py: Any
+    cm_to_qf_py: Any
     collinear_center_modes_py: Any
     compute_distance_series_py: Any
     compute_min_distance_py: Any
@@ -52,6 +53,7 @@ if TYPE_CHECKING:
     porkchop_grid_py: Any
     porkchop_grid_states_py: Any
     project_hamiltonian_qf_py: Any
+    qf_to_cm_py: Any
     qlaw_propagate_py: Any
     qlaw_segment_direction_py: Any
     propagate_bcr4bp_py: Any
@@ -108,6 +110,7 @@ _RUST_SYMBOLS = (
     "batch_synodic_to_j2000_py",
     "build_cr3bp_hamiltonian_py",
     "check_collision_py",
+    "cm_to_qf_py",
     "collinear_center_modes_py",
     "compute_distance_series_py",
     "compute_min_distance_py",
@@ -141,6 +144,7 @@ _RUST_SYMBOLS = (
     "porkchop_grid_py",
     "porkchop_grid_states_py",
     "project_hamiltonian_qf_py",
+    "qf_to_cm_py",
     "qlaw_propagate_py",
     "qlaw_segment_direction_py",
     "propagate_bcr4bp_py",
@@ -312,6 +316,7 @@ __all__ = [
     "batch_synodic_to_j2000_py",
     "build_cr3bp_hamiltonian_py",
     "check_collision_py",
+    "cm_to_qf_py",
     "collinear_center_modes_py",
     "CowellResult",
     "cowell_step",
@@ -354,6 +359,7 @@ __all__ = [
     "porkchop_grid_py",
     "porkchop_grid_states_py",
     "project_hamiltonian_qf_py",
+    "qf_to_cm_py",
     "qlaw_propagate_py",
     "qlaw_segment_direction_py",
     "propagate_compiled",